# internal/services/gateway/docs/swagger.yaml
basePath: /inn/api/v1
definitions:
  comment.deleteCommentRequest:
    properties:
      comment_id:
        type: integer
    type: object
  comment.makeCommentRequest:
    properties:
      avatar:
        type: string
      content:
        type: string
      from_user_id:
        type: string
      is_top:
        type: boolean
      post_id:
        type: integer
      title:
        type: string
      to_user_id:
        type: string
      top_comment_id:
        type: integer
    type: object
  internal.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  like.makeLikeRequest:
    properties:
      avatar:
        type: string
      post_id:
        type: integer
      title:
        type: string
    type: object
  post.createPostRequest:
    properties:
      category:
        properties:
          category_id:
            type: integer
          title:
            type: string
        type: object
      content:
        type: string
      tags:
        items:
          properties:
            tag:
              type: string
          type: object
        type: array
      title:
        type: string
    required:
    - category
    - content
    - title
    type: object
  post.deletePostRequest:
    properties:
      post_id:
        type: integer
    required:
    - post_id
    type: object
  post.getPostOfCategoryRequest:
    properties:
      category_id:
        type: integer
    required:
    - category_id
    type: object
  post.getPostOfTag:
    properties:
      title:
        type: string
    required:
    - title
    type: object
  post.searchRequest:
    properties:
      data:
        type: string
      search_type:
        type: string
    required:
    - data
    - search_type
    type: object
  user.user:
    properties:
      account:
        type: string
      password:
        type: string
    required:
    - account
    - password
    type: object
  user.userInfo:
    properties:
      account:
        type: string
      avatar:
        type: string
      nickname:
        type: string
      sex:
        type: integer
    type: object
host: 139.196.30.123
info:
  contact: {}
  description: Inn
  title: Inn
  version: "1.0"
paths:
  /post:
    delete:
      consumes:
      - application/json
      description: 删除帖子
      parameters:
      - description: 用户token
        in: header
        name: Authorzation
        required: true
        type: string
      - description: 帖子信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/post.deletePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 删除帖子 api
      tags:
      - post
    post:
      consumes:
      - application/json
      description: 创建帖子
      parameters:
      - description: 用户token
        in: header
        name: Authorzation
        required: true
        type: string
      - description: 帖子信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/post.createPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 创建帖子 api
      tags:
      - post
  /post/category:
    get:
      consumes:
      - application/json
      description: 获取分区帖子信息
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 条数
        in: query
        name: limit
        type: integer
      - description: 分类信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/post.getPostOfCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 获取分区帖子信息 api
      tags:
      - post
  /post/category/all:
    get:
      consumes:
      - application/json
      description: 获取分区信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 获取分区信息 api
      tags:
      - post
  /post/search:
    post:
      consumes:
      - application/json
      description: 搜索帖子
      parameters:
      - description: 搜索信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/post.searchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 搜索帖子 api
      tags:
      - post
  /post/tag:
    post:
      consumes:
      - application/json
      description: 获取tag下的帖子
      parameters:
      - description: tag信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/post.getPostOfTag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 获取tag下的帖子 api
      tags:
      - post
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户通过学号登录
      parameters:
      - description: 用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.user'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 登陆 api
      tags:
      - user
  /user/profile:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: 用户token
        in: header
        name: Authorzation
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 获取用户信息 api
      tags:
      - user
    put:
      consumes:
      - application/json
      description: 修改用户信息
      parameters:
      - description: 用户token
        in: header
        name: Authorzation
        required: true
        type: string
      - description: 需要修改的信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/user.userInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: 修改用户信息 api
      tags:
      - user
swagger: "2.0"